refactor: resolve MAC addresses at neigh discovery
This PR improves the efficiency of route resolution by eliminating the need for the link module.

Previously, we stored only basic neighbor information in the cache, and at the time of dataplane request formation, we had to resolve LinkIndex to MAC address. This was inefficient, as we were doing the resolution for every route formation request.

Changes:
- NeighbourEntry structure to store both next hop and interface MAC addresses
- Updated updateNeighbours() function in neigh.go.
  Now it resolves LinkIndex to MAC address during neighbor discovery.
  Added validation for MAC addresses (6 bytes, EUI-48 format)
- Removed dependency on the link module