refactor: use MAC address type instead of strings in protobuf
This commit changes the way we pass MAC addresses in protobuf. Instead
of strings we now pass them as a separate struct containing a single
64-bit field, which encodes a MAC address in big-endian.

This ensures that MAC addresses are correct, and no from-strings-parsing
in CLI and web components is now required. Previously we have to check
for errors while doing so.
Now we won't.